// src/dorc-web/src/apis/dorc-api/models/TerraformParameterType.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum TerraformParameterType {
    String = 'String',
    Number = 'Number',
    Bool = 'Bool',
    List = 'List',
    Map = 'Map',
    Object = 'Object'
}