// src/dorc-web/src/apis/dorc-api/models/ComponentApiModelTemplateApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    ComponentApiModel,
    ProjectApiModel,
} from './';

/**
 * @export
 * @interface ComponentApiModelTemplateApiModel
 */
export interface ComponentApiModelTemplateApiModel {
    /**
     * @type {ProjectApiModel}
     * @memberof ComponentApiModelTemplateApiModel
     */
    Project?: ProjectApiModel;
    /**
     * @type {Array<ComponentApiModel>}
     * @memberof ComponentApiModelTemplateApiModel
     */
    Items?: Array<ComponentApiModel> | null;
}